build: conditionalise output of .map files
If `NODE_ENV=production` then don't produce .map files for css and js
debugging.  This will allow for substantially smaller tarballs
(approximately half the size) and also decreases webpack a bit.

+++ pkg/lib/sassc-loader.js
@@ -4,6 +4,7 @@ const childProcess = require('child_process');
 
 const srcdir = process.env.SRCDIR || path.resolve(__dirname, '..', '..');
 const nodedir = path.resolve(srcdir, 'node_modules');
+const production = process.env.NODE_ENV === 'production';
 
 /* source is not used. This must be the first loader in the chain, using this.resource, so that sassc can include the scss file's directory in the include path */
 module.exports = function() {
@@ -20,15 +21,19 @@ module.exports = function() {
             '--load-path=' + path.resolve(nodedir, 'font-awesome-sass/assets/stylesheets'),
             '--load-path=' + path.resolve(nodedir, 'patternfly/dist/sass'),
             '--load-path=' + path.resolve(nodedir, 'bootstrap-sass/assets/stylesheets'),
-            '--style=compressed', '--sourcemap', this.resource, out
+            '--style=compressed', production ? '--' : '--sourcemap', this.resource, out
         ],
         { stdio: ['pipe', 'inherit', 'inherit'] });
 
     const css = fs.readFileSync(out, 'utf8');
-    const cssmap = fs.readFileSync(out + ".map", 'utf8');
-
     fs.unlinkSync(out);
-    fs.unlinkSync(out + ".map");
+
+    let cssmap;
+    if (!production) {
+        cssmap = fs.readFileSync(out + ".map", 'utf8');
+        fs.unlinkSync(out + ".map");
+    }
+
     fs.rmdirSync(workdir);
 
     this.callback(null, css, cssmap);
